min-requirements.txt
aioredis==2.0.1
celery==5.2.7
channels-redis==4.0.0
crispy-bootstrap5==0.7
daphne==4.0.0
Deprecated==1.2.13
dj-database-url==1.2.0
dj-email-url==1.0.6
django-bootstrap-v5==1.0.11
django-cache-url==3.4.4
django-redis==5.2.0
django-test-plus==2.2.1
djangorestframework==3.14.0
environs==9.5.0
hiredis==2.1.1
ipdb==0.13.11
model-bakery==1.9.0
mqttasgi==1.2.1
pip==22.0.4
pipdeptree==2.3.3
psycopg2-binary==2.9.5
py==1.11.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pytest-black==0.3.12
pytest-cov==4.0.0
pytest-django==4.5.2
requests==2.28.2
service-identity==21.1.0
wheel==0.37.1
whitenoise==6.3.0
